# Speed up `dftatom` execution time