--- a/src/libasr/pass/simplifier.cpp
+++ b/src/libasr/pass/simplifier.cpp
@@ -18,6 +18,9 @@ bool is_vectorise_able(ASR::expr_t* x) {
         case ASR::exprType::FunctionCall: {
             return ASRUtils::is_elemental(ASR::down_cast<ASR::FunctionCall_t>(x)->m_name);
         }
+        case ASR::exprType::IntrinsicElementalFunction: {
+            return true;
+        }
         case ASR::exprType::IntegerBinOp:
         case ASR::exprType::RealBinOp:
         case ASR::exprType::ComplexBinOp:
@@ -1136,8 +1139,7 @@ class ArgSimplifier: public ASR::CallReplacerOnExpressionsVisitor<ArgSimplifier>
     ASR::expr_t* visit_BinOp_expr(ASR::expr_t* expr, const std::string& name_hint, ASR::exprType allowed_expr) {
         if (ASRUtils::is_array(ASRUtils::expr_type(expr)) &&
             !ASR::is_a<ASR::ArrayBroadcast_t>(*expr) &&
-            !ASR::is_a<ASR::Var_t>(*ASRUtils::get_past_array_physical_cast(expr)) &&
-            !ASR::is_a<ASR::StructInstanceMember_t>(*ASRUtils::get_past_array_physical_cast(expr)) &&
+            !is_vectorise_able(expr) &&
             (expr->type != allowed_expr)
         ) {
             visit_expr(*expr);
@@ -1411,9 +1413,15 @@ class ArgSimplifier: public ASR::CallReplacerOnExpressionsVisitor<ArgSimplifier>
     void visit_ArrayBound(const ASR::ArrayBound_t& x) {
         ASR::ArrayBound_t& xx = const_cast<ASR::ArrayBound_t&>(x);
 
-        replace_expr_with_temporary_variable(v, "_array_bound_")
-        if (x.m_dim) {
-            replace_expr_with_temporary_variable(dim, "_array_bound_dim_")
+        if( is_temporary_needed(xx.m_v) ) {
+            replace_expr_with_temporary_variable(v, "_array_bound_v")
+        }
+    }
+
+    void visit_ArraySize(const ASR::ArraySize_t& x) {
+        ASR::ArraySize_t& xx = const_cast<ASR::ArraySize_t&>(x);
+        if( is_temporary_needed(xx.m_v) ) {
+            replace_expr_with_temporary_variable(v, "_array_size_v")
         }
     }
 };
@@ -1525,11 +1533,6 @@ class ReplaceExprWithTemporary: public ASR::BaseExprReplacer<ReplaceExprWithTemp
     void replace_IntrinsicImpureFunction(ASR::IntrinsicImpureFunction_t* x) {
         replace_current_expr(std::string("_intrinsic_impure_function_") +
             ASRUtils::get_impure_intrinsic_name(x->m_impure_intrinsic_id))
-    }
-
-    void replace_IntrinsicElementalFunction(ASR::IntrinsicElementalFunction_t* x) {
-        replace_current_expr(std::string("_intrinsic_elemental_function_") +
-            ASRUtils::get_intrinsic_name(x->m_intrinsic_id))
     }
 
     void replace_StructConstructor(ASR::StructConstructor_t* x) {
@@ -1743,8 +1746,7 @@ class ReplaceExprWithTemporary: public ASR::BaseExprReplacer<ReplaceExprWithTemp
     void replace_ArrayBound(ASR::ArrayBound_t* x) {
         ASR::expr_t** current_expr_copy_149 = current_expr;
         current_expr = &(x->m_v);
-        if( !ASR::is_a<ASR::Var_t>(*x->m_v) &&
-            !ASR::is_a<ASR::StructInstanceMember_t>(*x->m_v) ) {
+        if( is_temporary_needed(x->m_v) ) {
             force_replace_current_expr_for_array("_array_bound_v")
         }
         current_expr = current_expr_copy_149;
@@ -1753,8 +1755,7 @@ class ReplaceExprWithTemporary: public ASR::BaseExprReplacer<ReplaceExprWithTemp
     void replace_ArraySize(ASR::ArraySize_t* x) {
         ASR::expr_t** current_expr_copy_149 = current_expr;
         current_expr = &(x->m_v);
-        if( !ASR::is_a<ASR::Var_t>(*x->m_v) &&
-            !ASR::is_a<ASR::StructInstanceMember_t>(*x->m_v) ) {
+        if( is_temporary_needed(x->m_v) ) {
             force_replace_current_expr_for_array("_array_size_v")
         }
         current_expr = current_expr_copy_149;