Closes https://github.com/lfortran/lfortran/issues/5836

`./tests/atom_U/F_atom_U  0.76s user 0.01s system 99% cpu 0.774 total` (this PR)

`./tests/atom_U/F_atom_U  0.83s user 0.02s system 99% cpu 0.847 total` (with `--no-experimental-simplifier`)

Let's see what happens with tests. :D. Will rewrite git history before merging.